Add files via upload
This is a project I worked on to prompt users to enter a strong password. The program serves as a mechanism which the users to follow a set of security policies that will allow them to choose a strong password. In this case, the policies are as follows: choose a password that is at least 8 characters long; choose a password that contains at least one uppercase letter, at least one lowercase letter, at least one special character, and at least one numeric digit; the password cannot contain blank spaces; users cannot leave this field blank.

This project uses regular expressions (regex) to verify whether a password is strong or needs improvement.

This is a preliminary stage of the project, and next steps are to augment the code to include a scoring system to help users better understand how to improve their passwords.